build: bump flake.lock to fix OpenGL issues
Needed for GTK3 and OpenGL canvas, to prepare upcoming features of
showing screen previews in the GUI.

The error was that OpenGL complained with the error
"No GL implementation is available". By trial and error, it was
determined that updating the flake inputs made the go error away.

To determine the root cause, nixpkgs was bisected (which took multiple
days). The commit found was
https://redirect.github.com/NixOS/nixpkgs/commit/02828ac32324446552109118f78c35dfb3e8e494.
This suggests OpenGL apps requires libraries to be built off the exact
version of Wayland that the user's system is running on.